move sentry to it's own module
