+++ packages/modules/edge-api-sdk/package.json
@@ -28,10 +28,5 @@
     "@types/aws-lambda": "^8.10.121",
     "aws-lambda": "^1.0.7",
     "jest-fetch-mock": "^3.0.3"
-  },
-  "dependencies": {
-    "@sentry/node": "^7.109.0",
-    "@sentry/types": "^7.109.0",
-    "@sentry/utils": "^7.109.0"
   }
 }
