+++ packages/modules/sentry/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+  root: true,
+  extends: ["@reapit-cdk/eslint-config"],
+  parserOptions: {
+    ecmaVersion: 'ESNext',
+    sourceType: 'module',
+    project: './tsconfig.json',
+  },
+}
