+++ packages/modules/sentry/.npmignore
@@ -1,0 +1,2 @@
+tests
+.eslintrc.js
