+++ packages/modules/sentry/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@reapit-cdk/sentry",
+  "description": "Minimal Sentry SDK",
+  "main": "src/index.ts",
+  "version": "0.0.0",
+  "homepage": "https://github.com/reapit/ts-cdk-constructs/blob/main/packages/modules/sentry",
+  "readme": "https://github.com/reapit/ts-cdk-constructs/blob/main/packages/modules/sentry/readme.md",
+  "bugs": {
+    "url": "https://github.com/reapit/ts-cdk-constructs/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/reapit/ts-cdk-constructs.git"
+  },
+  "scripts": {
+    "publish": "yarn npm publish",
+    "check": "yarn run root:check -p $(pwd)",
+    "lint": "reapit-cdk-eslint",
+    "test": "yarn run root:test -- $(pwd)",
+    "build": "reapit-cdk-tsup",
+    "prepack": "yarn build"
+  },
+  "devDependencies": {
+    "@reapit-cdk/eslint-config": "workspace:^",
+    "@reapit-cdk/tsup": "workspace:^"
+  },
+  "dependencies": {
+    "@reapit-cdk/edge-api-sdk": "workspace:^",
+    "@sentry/node": "^7.112.2",
+    "@sentry/types": "^7.112.2",
+    "@sentry/utils": "^7.112.2"
+  }
+}
