+++ packages/modules/sentry/readme.md
@@ -1,0 +1,16 @@
+# @reapit-cdk/email-receiver-types
+
+
+![npm version](https://img.shields.io/npm/v/@reapit-cdk/email-receiver-types)
+![npm downloads](https://img.shields.io/npm/dm/@reapit-cdk/email-receiver-types)
+![coverage: 0%25](https://img.shields.io/badge/coverage-0%25-red)
+
+Types for @reapit-cdk/email-receiver and client.
+
+## Package Installation:
+
+```sh
+yarn add --dev @reapit-cdk/email-receiver-types
+# or
+npm install @reapit-cdk/email-receiver-types --save-dev
+```
