+++ packages/modules/sentry/src/edge-api-sentry-logger.ts
@@ -1,19 +1,22 @@
-import { createGetModuleFromFilename } from './module'
-
-import { createEventEnvelope, getEnvelopeEndpointWithUrlEncodedAuth } from '@sentry/core'
+import { SeverityLevel } from '@sentry/types'
 import {
   stackParserFromStackParserOptions,
   createStackParser,
   nodeStackLineParser,
   exceptionFromError,
-  dsnFromString,
 } from '@sentry/utils'
+import {
+  RCQuery,
+  RCHeaders,
+  RCRequest,
+  LogLevel,
+  Transport,
+  MinimalLogPayload,
+  LogPayload,
+} from '@reapit-cdk/edge-api-sdk'
 
-import { SeverityLevel } from '@sentry/types'
-
-import { LogLevel, LogPayload, MinimalLogPayload, Transport } from '../logger'
-import { RCHeaders, RCQuery, RCRequest } from '../types'
-import { serializeEnvelope } from './serialize-envelope'
+import { createGetModuleFromFilename } from './module'
+import { getCookies, sendEnvelope, validateSentryDsn } from './sentry'
 
 const stackParser = stackParserFromStackParserOptions(
   createStackParser(nodeStackLineParser(createGetModuleFromFilename())),
@@ -26,34 +29,20 @@ const logLevelToSeverityLevel = (logLevel: LogLevel): SeverityLevel => {
   return logLevel
 }
 
-export const getCookies = (cookie: string): Record<string, string> =>
-  Object.fromEntries(
-    cookie
-      .split('; ')
-      .map((v) => {
-        try {
-          return v.split(/=(.*)/s).map(decodeURIComponent)
-        } catch {
-          return ['', '']
-        }
-      })
-      .filter(([v]) => !!v),
-  )
-
-const pickFirst = (queryOrHeaders: RCQuery | RCHeaders) => {
-  return Object.entries(queryOrHeaders)
+const pickFirst = (queryOrHeaders: RCQuery | RCHeaders) =>
+  Object.entries(queryOrHeaders)
     .map(([k, v]) => [k, Array.isArray(v) ? v[0] : v])
-    .reduce((pv, cv) => {
-      return {
+    .reduce(
+      (pv, cv) => ({
         ...pv,
         [cv[0]]: cv[1],
-      }
-    }, {})
-}
+      }),
+      {},
+    )
 
 const dateToSecs = (d: Date) => Math.round(d.getTime() / 1000)
 
-export const init = (request: RCRequest<{ sentryDsn: string; sentryRelease: string }>): Transport => {
+export const initSentryLogger = (request: RCRequest<{ sentryDsn: string; sentryRelease: string }>): Transport => {
   const {
     host,
     cookies,
@@ -61,15 +50,14 @@ export const init = (request: RCRequest<{ sentryDsn: string; sentryRelease: stri
     region,
     meta: { sessionId, functionName, functionVersion, invocationId },
   } = request
-  const components = dsnFromString(sentryDsn)
-  if (!components) {
-    throw new Error('invalid DSN provided')
-  }
+
+  validateSentryDsn(sentryDsn)
 
   return async (payload: MinimalLogPayload | LogPayload) => {
     const errorEntry = payload.entries.find((entry) => !!entry.error)
     const exception = errorEntry?.error ? exceptionFromError(stackParser, errorEntry.error) : undefined
-    const envelope = createEventEnvelope({
+
+    await sendEnvelope(sentryDsn, {
       exception: exception
         ? {
             values: [exception],
@@ -93,31 +81,21 @@ export const init = (request: RCRequest<{ sentryDsn: string; sentryRelease: stri
         method: request.method,
         query_string: request.query ? pickFirst(request.query) : undefined,
         cookies: cookies.map(getCookies).reduce(
-          (pv, cv) => {
-            return {
-              ...pv,
-              ...cv,
-            }
-          },
+          (pv, cv) => ({
+            ...pv,
+            ...cv,
+          }),
           {} as Record<string, string>,
         ),
         url: request.path,
       },
       release: sentryRelease,
       environment: host,
-      breadcrumbs: payload.entries.map((entry) => {
-        return {
-          level: logLevelToSeverityLevel(entry.level),
-          message: entry.message,
-          timestamp: dateToSecs(entry.timestamp),
-        }
-      }),
-    })
-
-    const endpoint = getEnvelopeEndpointWithUrlEncodedAuth(components)
-    await fetch(endpoint, {
-      body: serializeEnvelope(envelope),
-      method: 'post',
+      breadcrumbs: payload.entries.map((entry) => ({
+        level: logLevelToSeverityLevel(entry.level),
+        message: entry.message,
+        timestamp: dateToSecs(entry.timestamp),
+      })),
     })
   }
 }
