+++ packages/modules/sentry/src/index.ts
@@ -1,0 +1,2 @@
+export * from './edge-api-sentry-logger'
+export * from './sentry'
