+++ packages/modules/sentry/src/module.ts
