+++ packages/modules/sentry/src/serialize-envelope.ts
